fix: инициализация происходит в одном файле
