--- frontend/src/init.jsx
@@ -1,30 +1,0 @@
-import i18next from 'i18next';
-import React from 'react';
-import { I18nextProvider, initReactI18next } from 'react-i18next';
-
-import App from './App';
-import resources from './locales/index.js';
-import './locales/profanityList.js';
-
-const init = async () => {
-  const i18n = i18next.createInstance();
-
-  await i18n
-    .use(initReactI18next)
-    .init({
-      resources,
-      debug: true,
-      fallbackLng: 'ru',
-      interpolation: {
-        escapeValue: false,
-      },
-    });
-
-  return (
-    <I18nextProvider i18n={i18n}>
-      <App />
-    </I18nextProvider>
-  );
-};
-
-export default init;
